tests/wpt/web-platform-tests/css/css-animations/parsing/animation-range-end-computed.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#propdef-animation-range">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target" style="font-size:10px;"></div>
<script>
test_computed_value("animation-range-end", "initial", "auto");
test_computed_value("animation-range-end", "auto");
test_computed_value("animation-range-end", "cover 0%");
test_computed_value("animation-range-end", "COVER 0%", "cover 0%");
test_computed_value("animation-range-end", "cover 100%");
test_computed_value("animation-range-end", "cover 120%");
test_computed_value("animation-range-end", "cover 42%");
test_computed_value("animation-range-end", "cover -42%");
test_computed_value("animation-range-end", "contain 42%");
test_computed_value("animation-range-end", "exit 42%");
test_computed_value("animation-range-end", "exit calc(41% + 1%)", "exit 42%");
test_computed_value("animation-range-end", "exit 1%, cover 2%, contain 100%");
test_computed_value("animation-range-end", "enter 42px");
test_computed_value("animation-range-end", "contain calc(10% + 10px)");
test_computed_value("animation-range-end", "enter 1em", "enter 10px");
test_computed_value("animation-range-end", "exit calc(1em + 10px)", "exit 20px");
</script>
